refactor: move skip-teardown fixture to shared conftest
Move _skip_teardown_if_requested from autorag conftest to the parent
pipelines_components conftest so it applies to both AutoML and AutoRAG
tests. Rename env var from AUTORAG_SKIP_TEARDOWN to SKIP_TEARDOWN.
