+++ tests/pipelines_components/autorag/conftest.py
@@ -183,26 +183,6 @@ def _validate_autorag_env() -> None:
         pytest.skip("AutoRAG smoke test requires environment variables:\n" + "\n".join(missing))
 
 
-@pytest.fixture(scope="session", autouse=True)
-def _skip_teardown_if_requested():
-    """When AUTORAG_SKIP_TEARDOWN=true, prevent all ocp_resources from being deleted on teardown.
-
-    This keeps the namespace, pods, routes, and all resources alive after the test
-    so that other teams can inspect logs and state.
-    Delete the namespace manually when done: oc delete ns <namespace>
-    """
-    if os.getenv("AUTORAG_SKIP_TEARDOWN", "").lower() not in ("true", "1", "yes"):
-        yield
-        return
-    LOGGER.warning("AUTORAG_SKIP_TEARDOWN is set — all resources will be kept after test completion")
-    from ocp_resources.resource import Resource
-
-    _original_clean_up = Resource.clean_up
-    Resource.clean_up = lambda self, *args, **kwargs: True
-    yield
-    Resource.clean_up = _original_clean_up
-
-
 # ---------------------------------------------------------------------------
 # Override parent namespace fixture with a shorter name to stay under the
 # 63-char DNS label limit for KServe predictor hostnames.
